create empty type and import it onto notification component
