+++ src/components/Notifications/Notifications.tsx
@@ -1,21 +1,33 @@
-import { useAppSelector, useNotification } from '../../utilities/redux/hooks';
-import { useSelector } from 'react-redux';
-import { RootState } from '../../utilities/redux/store';
-import { Snackbar, Alert, SnackbarCloseReason } from '@mui/material';
+import { useState, useEffect } from 'react';
+import { io } from 'socket.io-client';
+import { emptyNotification } from '../../utilities/data/notificationConstants';
+import { NotificationState } from '../../utilities/types/NotificationInterface';
+const ENDPOINT = `${process.env.REACT_APP_LOCAL_URL}`;
+const socket = io(ENDPOINT, { transports: ['websocket'] });
 
-export const Notifications = (): JSX.Element => {
-  const notification = useSelector((state: RootState) => state.notification);
-  const { clearNotification } = useNotification();
+export const Notifications = () => {
+  const [notifications, setNotifications] = useState<NotificationState>(emptyNotification);
 
-  const handleClose = (_: unknown, reason?: SnackbarCloseReason) => {
-    reason !== 'clickaway' && clearNotification();
-  };
+  useEffect(() => {
+    // const askUserPermission = async () => {
+    //   return await Notification.requestPermission();
+    // };
+    // askUserPermission();
+
+    // const handler = (notifications: string) => {
+    //   setNotifications([...notifications, notifications]);
+    // };
+
+    socket.on('changes', () => {
+      console.log(socket.connected);
+    });
+  }, []);
+
+  console.log(notifications);
 
   return (
-    <Snackbar open={notification.open} autoHideDuration={notification.timeout} onClose={handleClose}>
-      <Alert variant="filled" onClose={handleClose} severity={notification.type}>
-        {notification.message}
-      </Alert>
-    </Snackbar>
+    <>
+      <p>hi</p>
+    </>
   );
 };
