+++ src/index.tsx
@@ -6,13 +6,13 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import reportWebVitals from './reportWebVitals';
 import { Provider } from 'react-redux';
 import { store } from './utilities/redux/store';
-import { Notifications } from './components/Notifications/Notifications';
+import { NotificationAlerts } from './components/Notifications/NotificationAlerts';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
   <Router>
     <Provider store={store}>
-      <Notifications />
+      <NotificationAlerts />
       <App />
     </Provider>
   </Router>,
