+++ src/utilities/data/notificationConstants.ts
@@ -1,8 +1,16 @@
-import { NotificationState } from '../types/NotificationInterface';
+import { NotificationInterface, NotificationState } from '../types/NotificationInterface';
 
-export const notificationInitialState: NotificationState = {
+export const notificationInitialState: NotificationInterface = {
   open: false,
   type: 'info',
   message: '',
   timeout: 5000,
+};
+
+export const emptyNotification: NotificationState = {
+  user: '',
+  notification: '',
+  message: '',
+  read: false,
+  _id: '',
 };
