+++ src/utilities/types/NotificationInterface.ts
@@ -1,9 +1,16 @@
 import { AlertColor } from '@mui/material';
-import { createSlice, createAsyncThunk, PayloadAction } from '@reduxjs/toolkit';
 
-export interface NotificationState {
+export interface NotificationInterface {
   open?: boolean;
   type?: AlertColor;
   message?: string;
   timeout?: number | null;
+}
+
+export interface NotificationState {
+  user: string;
+  notification: string;
+  message: string;
+  read: boolean;
+  _id: string;
 }
